requirements.txt
discord.py==2.3.2
aiosqlite==3.0.0